refactor: rename example migration to match timestamped convention
unimplemented_create_table was misleading since it's fully implemented.
Renamed to m20220101_000001_create_table to match the mYYYYMMDD_HHMMSS_*
naming convention used by the generator.
