+++ examples/migration/src/lib.rs
@@ -1,4 +1,4 @@
-mod unimplemented_create_table;
+mod m20220101_000001_create_table;
 
 use sea_orm_migration_spanner::prelude::*;
 
@@ -6,6 +6,6 @@ pub struct Migrator;
 
 impl MigratorTrait for Migrator {
     fn migrations() -> Vec<Box<dyn MigrationTrait>> {
-        vec![Box::new(unimplemented_create_table::Migration)]
+        vec![Box::new(m20220101_000001_create_table::Migration)]
     }
 }
